+++ tests.c
@@ -682,33 +682,33 @@ TEST(to_c_str) {
     assert(strcmp(Slice99_c_str(slice, (char[4]){0}), "abc") == 0);
 }
 
-TEST(pack_to_u8) {
-    uint8_t n = UINT8_MAX;
-    Slice99 slice = Slice99_from_array((uint8_t[]){n});
+#define TEST_PACK_TO_Ux(x)                                                                         \
+    TEST(pack_to_u##x##_beginning) {                                                               \
+        uint##x##_t n = UINT##x##_MAX;                                                             \
+        Slice99 slice = Slice99_from_array((uint##x##_t[]){n});                                    \
+        assert(Slice99_pack_to_u##x##_beginning(slice) == n);                                      \
+    }
 
-    assert(Slice99_pack_to_u8(slice) == n);
-}
+TEST_PACK_TO_Ux(8)
+TEST_PACK_TO_Ux(16)
+TEST_PACK_TO_Ux(32)
+TEST_PACK_TO_Ux(64)
 
-TEST(pack_to_u16) {
-    uint16_t n = UINT16_MAX;
-    Slice99 slice = Slice99_from_array((uint16_t[]){n});
+#undef TEST_PACK_TO_Ux
 
-    assert(Slice99_pack_to_u16(slice) == n);
-}
+#define TEST_PACK_TO_Ux_BEGINNING(x)                                                               \
+    TEST(pack_to_u##x##_beginning) {                                                               \
+        uint##x##_t n = UINT##x##_MAX;                                                             \
+        Slice99 slice = Slice99_from_array((uint##x##_t[]){n, 123, 456});                          \
+        assert(Slice99_pack_to_u##x##_beginning(slice) == n);                                      \
+    }
 
-TEST(pack_to_u32) {
-    uint32_t n = UINT32_MAX;
-    Slice99 slice = Slice99_from_array((uint32_t[]){n});
+TEST_PACK_TO_Ux_BEGINNING(8)
+TEST_PACK_TO_Ux_BEGINNING(16)
+TEST_PACK_TO_Ux_BEGINNING(32)
+TEST_PACK_TO_Ux_BEGINNING(64)
 
-    assert(Slice99_pack_to_u32(slice) == n);
-}
-
-TEST(pack_to_u64) {
-    uint64_t n = UINT64_MAX;
-    Slice99 slice = Slice99_from_array((uint64_t[]){n});
-
-    assert(Slice99_pack_to_u64(slice) == n);
-}
+#undef TEST_PACK_TO_Ux_BEGINNING
 
 // Typed slice {
 
@@ -833,6 +833,11 @@ int main(void) {
     test_pack_to_u16();
     test_pack_to_u32();
     test_pack_to_u64();
+
+    test_pack_to_u8_beginning();
+    test_pack_to_u16_beginning();
+    test_pack_to_u32_beginning();
+    test_pack_to_u64_beginning();
 
     test_typed_slice();
     test_fundamentals();